move everything test-related into test subfolder
