+++ src/ystdlib/error_handling/CMakeLists.txt
@@ -2,8 +2,8 @@ cpp_library(
     NAME error_handling
     NAMESPACE ystdlib
     TESTS_SOURCES
-        test-constants.hpp
-        test-types.hpp
-        test-types.cpp
-        test-ErrorCode.cpp
+        test/constants.hpp
+        test/types.hpp
+        test/types.cpp
+        test/test_error_code.cpp
 )
